+++ packages/camera/camera_platform_interface/test/camera_platform_interface_test.dart
@@ -109,6 +109,32 @@ void main() {
       );
     });
 
+    test(
+        'Default implementation of lockCaptureOrientation() should throw unimplemented error',
+        () {
+      // Arrange
+      final cameraPlatform = ExtendsCameraPlatform();
+
+      // Act & Assert
+      expect(
+        () => cameraPlatform.lockCaptureOrientation(1, null),
+        throwsUnimplementedError,
+      );
+    });
+
+    test(
+        'Default implementation of unlockCaptureOrientation() should throw unimplemented error',
+        () {
+      // Arrange
+      final cameraPlatform = ExtendsCameraPlatform();
+
+      // Act & Assert
+      expect(
+        () => cameraPlatform.unlockCaptureOrientation(1),
+        throwsUnimplementedError,
+      );
+    });
+
     test('Default implementation of dispose() should throw unimplemented error',
         () {
       // Arrange