+++ packages/camera/camera_platform_interface/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.4.0
+
+- Introduces interface methods for locking and unlocking the capture orientation.
+- Introduces interface method for listening to the device orientation.
+
 ## 1.3.0
 
 - Introduces an option to set the image format when initializing.